requirements.txt
easyocr
opencv-python
matplotlib
numpy
streamlit
transformers
torch
torchvision
pillow
sentencepiece
